% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gut_IEU.R
\name{Gut_IEU}
\alias{Gut_IEU}
\title{肠道菌群MR工具}
\usage{
Gut_IEU(savefile, PATH, GWASID, outname, kb, r2)
}
\arguments{
\item{savefile}{设置一个保存结果的文件夹名字}

\item{PATH}{切分的肠道菌群暴露文件位置}

\item{GWASID}{结局的GWAS id号}

\item{outname}{结局的英文名称或者缩写}

\item{kb}{聚类的距离}

\item{r2}{相关系数}
}
\description{
肠道菌群MR工具
}
