% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gut_local.R
\name{Gut_local}
\alias{Gut_local}
\title{肠道菌群MR工具}
\usage{
Gut_local(name, key, savefile, PATH, GWASsummay, outname, kb, r2)
}
\arguments{
\item{name}{学号}

\item{key}{密码}

\item{savefile}{设置一个保存结果的文件夹名字}

\item{PATH}{切分的肠道菌群暴露文件位置}

\item{GWASsummay}{本地结局的预处理好的GWAS summay}

\item{outname}{本地结局的名称}

\item{kb}{聚类距离}

\item{r2}{相关系数}
}
\description{
肠道菌群MR工具
}
