% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IEU_Gut.R
\name{IEU_Gut}
\alias{IEU_Gut}
\title{肠道菌群MR工具}
\usage{
IEU_Gut(name, key, savefile, PATH, GWASID, expname, p1, r2, kb)
}
\arguments{
\item{name}{学号}

\item{key}{密码}

\item{savefile}{设置一个保存结果的文件夹名字}

\item{PATH}{GUT_outcome文件位置}

\item{GWASID}{暴露的GWAS id号}

\item{expname}{暴露的名称}

\item{p1}{阈值}

\item{r2}{相关系数}

\item{kb}{聚类距离}
}
\description{
肠道菌群MR工具
}
