% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Local_Gut.R
\name{Local_Gut}
\alias{Local_Gut}
\title{肠道菌群MR工具}
\usage{
Local_Gut(name, key, savefile, PATH, GWASIV, expname)
}
\arguments{
\item{name}{学号}

\item{key}{密码}

\item{savefile}{设置一个保存结果的文件夹名字}

\item{PATH}{GUT_outcome文件位置}

\item{GWASIV}{本地预处理好的暴露IV}

\item{expname}{暴露的名称}
}
\description{
肠道菌群MR工具
}
